Restructuring of classes.
