+++ app/src/main/java/com/ahujafabrics/yarnit/Adapter/ViewOrderItem.java
@@ -1,14 +1,14 @@
-package com.ahujafabrics.yarnit.Activity;
+package com.ahujafabrics.yarnit.Adapter;
 
 import android.content.Context;
 import android.content.Intent;
-import android.os.Parcelable;
 import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.TextView;
 
+import com.ahujafabrics.yarnit.Activity.ViewOrderDetails;
 import com.ahujafabrics.yarnit.R;
 import com.ahujafabrics.yarnit.Repository.Order;
