--- app/src/main/java/com/ahujafabrics/yarnit/Alerts/CreateProfileAlert.java
@@ -1,26 +1,0 @@
-package com.ahujafabrics.yarnit.Alerts;
-
-import android.app.Dialog;
-import android.content.DialogInterface;
-import android.os.Bundle;
-import android.support.v4.app.DialogFragment;
-import android.support.v7.app.AlertDialog;
-
-import com.ahujafabrics.yarnit.R;
-
-public class CreateProfileAlert extends DialogFragment {
-    @Override
-    public Dialog onCreateDialog(Bundle savedInstanceState) {
-        // Use the Builder class for convenient dialog construction
-        AlertDialog.Builder builder = new AlertDialog.Builder(getActivity());
-        builder.setMessage(R.string.NoProfileAlert)
-                .setPositiveButton(R.string.Ok, new DialogInterface.OnClickListener() {
-                    public void onClick(DialogInterface dialog, int id) {
-                        dialog.dismiss();
-                    }
-                });
-
-        // Create the AlertDialog object and return it
-        return builder.create();
-    }
-}
