+++ app/src/main/res/layout/activity_catalog.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Catalog"
+    tools:context=".Activity.Catalog"
     android:gravity="center_horizontal"
     android:orientation="vertical">
