+++ app/src/main/res/layout/activity_dashboard.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Dashboard"
+    tools:context=".Activity.Dashboard"
     android:orientation="vertical"
     android:padding="10dp"
     android:background="@color/backgroundcolor"
