+++ app/src/main/res/layout/activity_profile.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".Profile">
+    tools:context=".Activity.Profile">
 
     <View
         android:id="@+id/view"
